# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KathArtsis Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface Transaction {
  id: number;
  date: string;
  type: TransactionType;
  category: string;
  name: string;
  amount: number;
  paymentMethod: string;
  notes?: string;
  createdAt?: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateTransactionBody {
  date: string;
  type: CreateTransactionBodyType;
  category: string;
  name: string;
  amount: number;
  paymentMethod: string;
  notes?: string;
}

export interface FinanceSummary {
  totalIncome: number;
  totalExpense: number;
  balance: number;
}

export interface Receipt {
  id: number;
  receiptNumber: string;
  donorName: string;
  amount: number;
  paymentMethod: string;
  createdAt: string;
}

export interface CreateReceiptBody {
  donorName: string;
  amount: number;
  paymentMethod: string;
}

export type StickerPosition =
  (typeof StickerPosition)[keyof typeof StickerPosition];

export const StickerPosition = {
  "1st": "1st",
  "2nd": "2nd",
  "3rd": "3rd",
} as const;

export interface Sticker {
  id: number;
  programName: string;
  position: StickerPosition;
  name: string;
  createdAt: string;
}

export type CreateStickerBodyPosition =
  (typeof CreateStickerBodyPosition)[keyof typeof CreateStickerBodyPosition];

export const CreateStickerBodyPosition = {
  "1st": "1st",
  "2nd": "2nd",
  "3rd": "3rd",
} as const;

export interface CreateStickerBody {
  programName: string;
  position: CreateStickerBodyPosition;
  name: string;
}

export interface Certificate {
  id: number;
  name: string;
  programName: string;
  position: string;
  createdAt: string;
}

export interface CreateCertificateBody {
  name: string;
  programName: string;
  position: string;
}