# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KathArtsis Manager API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all transactions
 */
export const ListTransactionsResponseItem = zod.object({
  id: zod.number(),
  date: zod.string(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  name: zod.string(),
  amount: zod.number(),
  paymentMethod: zod.string(),
  notes: zod.string().optional(),
  createdAt: zod.string().optional(),
});
export const ListTransactionsResponse = zod.array(ListTransactionsResponseItem);

/**
 * @summary Create a transaction
 */
export const CreateTransactionBody = zod.object({
  date: zod.string(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  name: zod.string(),
  amount: zod.number(),
  paymentMethod: zod.string(),
  notes: zod.string().optional(),
});

/**
 * @summary Delete a transaction
 */
export const DeleteTransactionParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteTransactionResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get finance summary totals
 */
export const GetFinanceSummaryResponse = zod.object({
  totalIncome: zod.number(),
  totalExpense: zod.number(),
  balance: zod.number(),
});

/**
 * @summary List all receipts
 */
export const ListReceiptsResponseItem = zod.object({
  id: zod.number(),
  receiptNumber: zod.string(),
  donorName: zod.string(),
  amount: zod.number(),
  paymentMethod: zod.string(),
  createdAt: zod.string(),
});
export const ListReceiptsResponse = zod.array(ListReceiptsResponseItem);

/**
 * @summary Create a donation receipt
 */
export const CreateReceiptBody = zod.object({
  donorName: zod.string(),
  amount: zod.number(),
  paymentMethod: zod.string(),
});

/**
 * @summary List all stickers
 */
export const ListStickersResponseItem = zod.object({
  id: zod.number(),
  programName: zod.string(),
  position: zod.enum(["1st", "2nd", "3rd"]),
  name: zod.string(),
  createdAt: zod.string(),
});
export const ListStickersResponse = zod.array(ListStickersResponseItem);

/**
 * @summary Create a sticker
 */
export const CreateStickerBody = zod.object({
  programName: zod.string(),
  position: zod.enum(["1st", "2nd", "3rd"]),
  name: zod.string(),
});

/**
 * @summary Delete a sticker
 */
export const DeleteStickerParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteStickerResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary List all certificates
 */
export const ListCertificatesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  programName: zod.string(),
  position: zod.string(),
  createdAt: zod.string(),
});
export const ListCertificatesResponse = zod.array(ListCertificatesResponseItem);

/**
 * @summary Create a certificate
 */
export const CreateCertificateBody = zod.object({
  name: zod.string(),
  programName: zod.string(),
  position: zod.string(),
});

/**
 * @summary Delete a certificate
 */
export const DeleteCertificateParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteCertificateResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});