# lib/api-zod/src/generated/types/createCertificateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KathArtsis Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCertificateBody {
  name: string;
  programName: string;
  position: string;
}