# lib/api-zod/src/generated/types/createStickerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KathArtsis Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateStickerBodyPosition } from "./createStickerBodyPosition";

export interface CreateStickerBody {
  programName: string;
  position: CreateStickerBodyPosition;
  name: string;
}