# lib/api-zod/src/generated/types/createTransactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KathArtsis Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionBodyType } from "./createTransactionBodyType";

export interface CreateTransactionBody {
  date: string;
  type: CreateTransactionBodyType;
  category: string;
  name: string;
  amount: number;
  paymentMethod: string;
  notes?: string;
}