# lib/api-zod/src/generated/types/financeSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KathArtsis Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface FinanceSummary {
  totalIncome: number;
  totalExpense: number;
  balance: number;
}