# lib/api-zod/src/generated/types/receipt.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KathArtsis Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Receipt {
  id: number;
  receiptNumber: string;
  donorName: string;
  amount: number;
  paymentMethod: string;
  createdAt: string;
}